animals/requirements.txt
asgiref==3.7.2
boto3==1.28.71
botocore==1.31.71
certifi==2023.7.22
Django==4.2.5
django-storages==1.14.2
djangorestframework==3.14.0
jmespath==1.0.1
minio==7.1.17
Pillow==10.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
